feat:채널 조회 기능 구현

+++ src/main/java/com/bipa4/back_bipatv/entity/Channels.java
@@ -18,7 +18,7 @@ public class Channels {
   @Id
   @GeneratedValue(strategy = GenerationType.IDENTITY)
   @Column(name = "channel_id")
-  private int channelId;
+  private Long channelId;
   @Column(name = "name", nullable = false)
   private String name;
   @Column(name = "content", nullable = true)
